don't use mimetypes
Associating it by extension only to then match the mimetype is just adding extra steps. If we're already using a match for the
extension, might as well just call the logic from that than convert it to mimetype *and then* optimize based on that.

Also the status of the "mime" crate is a bit uncertain at the moment, so depending on it this early in the project isn't a great
call